refactor: register msg service
