--- x/blob/handler.go
@@ -1,27 +1,0 @@
-package blob
-
-import (
-	"fmt"
-
-	"github.com/celestiaorg/celestia-app/x/blob/keeper"
-	"github.com/celestiaorg/celestia-app/x/blob/types"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-)
-
-// NewHandler uses the provided blob keeper to create an sdk.Handler
-func NewHandler(k keeper.Keeper) sdk.Handler {
-	msgServer := keeper.NewMsgServerImpl(k)
-
-	return func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
-		ctx = ctx.WithEventManager(sdk.NewEventManager())
-		switch msg := msg.(type) {
-		case *types.MsgPayForBlob:
-			res, err := msgServer.PayForBlob(sdk.WrapSDKContext(ctx), msg)
-			return sdk.WrapServiceResult(ctx, res, err)
-		default:
-			errMsg := fmt.Sprintf("unrecognized %s message type: %T", types.ModuleName, msg)
-			return nil, sdkerrors.Wrap(sdkerrors.ErrUnknownRequest, errMsg)
-		}
-	}
-}
